requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
